requirements.txt
decord
numpy
pillow
openai
torch